+++ packages/job-worker/src/playout/__tests__/resolvedPieces.test.ts
@@ -5,7 +5,7 @@ import { MockJobContext, setupDefaultJobEnvironment } from '../../__mocks__/cont
 import { SourceLayers } from '@sofie-automation/corelib/dist/dataModel/ShowStyleBase'
 import { ReadonlyDeep } from 'type-fest'
 import { protectString } from '@sofie-automation/shared-lib/dist/lib/protectedString'
-import { getRandomId, normalizeArrayToMap } from '@sofie-automation/corelib/dist/lib'
+import { getRandomId } from '@sofie-automation/corelib/dist/lib'
 import { IBlueprintPieceType, PieceLifespan } from '@sofie-automation/blueprints-integration'
 import {
 	PieceInstance,
@@ -15,10 +15,8 @@ import {
 import { EmptyPieceTimelineObjectsBlob } from '@sofie-automation/corelib/dist/dataModel/Piece'
 import _ = require('underscore')
 import { JobContext } from '../../jobs'
-import {
-	PieceInstanceWithTimings,
-	processAndPrunePieceInstanceTimings,
-} from '@sofie-automation/corelib/dist/playout/processAndPrune'
+import { processAndPrunePieceInstanceTimings } from '@sofie-automation/corelib/dist/playout/processAndPrune'
+import { resolvePrunedPieceInstances } from '../resolvedPieces'
 
 describe('Resolved Pieces', () => {
 	let context: MockJobContext
@@ -73,53 +71,6 @@ describe('Resolved Pieces', () => {
 			},
 			userDuration: instancePartial?.userDuration,
 		}
-	}
-
-	function resolveStartOfInstance(nowInPart: number, instance: PieceInstanceWithTimings): number {
-		return instance.piece.enable.start === 'now' ? nowInPart ?? 0 : instance.piece.enable.start
-	}
-
-	function resolvePrunedPieceInstances(
-		nowInPart: number,
-		pieceInstances: PieceInstanceWithTimings[]
-	): ResolvedPieceInstance[] {
-		const pieceInstancesMap = normalizeArrayToMap(pieceInstances, '_id')
-
-		// Calculate the durations
-		return pieceInstances.map((instance) => {
-			const resolvedStart = resolveStartOfInstance(nowInPart, instance)
-
-			let resolvedEnd: number | undefined
-			if (typeof instance.resolvedEndCap === 'number') {
-				resolvedEnd = instance.resolvedEndCap
-				// } else if (instance.resolvedEndCap === 'now') {
-				// 	// TODO - something should test this route?
-				// 	// resolvedEnd = nowInPart
-			} else if (instance.resolvedEndCap) {
-				const otherInstance = pieceInstancesMap.get(instance.resolvedEndCap.relativeToStartOf)
-				if (otherInstance) {
-					resolvedEnd = resolveStartOfInstance(nowInPart, otherInstance)
-				}
-			}
-
-			const caps: number[] = []
-			if (resolvedEnd !== undefined) caps.push(resolvedEnd - resolvedStart)
-			if (instance.piece.enable.duration !== undefined) caps.push(instance.piece.enable.duration)
-
-			if (instance.userDuration) {
-				if ('endRelativeToPart' in instance.userDuration) {
-					caps.push(instance.userDuration.endRelativeToPart - resolvedStart)
-				} else if ('endRelativeToNow' in instance.userDuration) {
-					caps.push(nowInPart + instance.userDuration.endRelativeToNow - resolvedStart)
-				}
-			}
-
-			return {
-				...instance,
-				resolvedStart,
-				resolvedDuration: caps.length ? Math.min(...caps) : undefined,
-			}
-		})
 	}
 
 	function getResolvedPiecesInner(
